% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-pkg.R
\name{build_pkg}
\alias{build_pkg}
\alias{get_build_status}
\alias{get_build_manifest}
\alias{start_build}
\title{Submits a build job for a package}
\usage{
build_pkg(package = ".", note = "New build", tags = c("package",
  "rstats"), read_access = NULL, write_access = NULL, execute = TRUE,
  secrets = TRUE, browse = FALSE, username = sourcehut_user(),
  pat = sourcehut_pat())

get_build_status(job_id, pat = sourcehut_pat())

get_build_manifest(job_id, pat = sourcehut_pat())

start_build(job_id, pat = sourcehut_pat())
}
\arguments{
\item{package}{path to package (assumes current directory)}

\item{note}{human-friendly description of this build (markdown, optional)}

\item{tags}{character vector of strings that identify this build and can be
used to navigate the dashboard. Each string must use only lowercase
alphanumeric characters, or any of "-_."}

\item{read_access, write_access}{character vector of users that have read or write access to the job.}

\item{execute}{start build immediately? (Default: \code{TRUE})}

\item{secrets}{provide secrets to the build? (Default: \code{TRUE})}

\item{browse}{open browser to the builder URL? (Default: \code{FALSE})}

\item{username}{see \code{\link[=sourcehut_user]{sourcehut_user()}}}

\item{pat}{see \code{\link[=sourcehut_pat]{sourcehut_pat()}}}

\item{job_id}{\code{builds.sr.ht} job id}
}
\description{
This function assumes you know what you're doing and expects a valid
\code{.build.yml} in the \code{package} directory.
}
\references{
\url{https://man.sr.ht/builds.sr.ht/api.md}
}
