% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_repos}
\alias{git_repos}
\title{Retrieve metadata about repositories}
\usage{
git_repos(username = NULL, next_id = 1, pat = sourcehut_pat())
}
\arguments{
\item{username}{if not \code{NULL} then a valid SourceHut username}

\item{next_id}{if there is a \code{next} field in the API response and you wish
to paginate through the results, feed the value of \code{next} to
the \code{next_id} parameter.}

\item{pat}{see \code{\link[=sourcehut_pat]{sourcehut_pat()}}}
}
\description{
Retrieve metadata about repositories
}
\references{
\url{https://man.sr.ht/git.sr.ht/api.md}
}
