% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_user}
\alias{git_user}
\title{Retrieve metadata about yourself or another SourceHut user}
\usage{
git_user(username = NULL, pat = sourcehut_pat())
}
\arguments{
\item{username}{if not \code{NULL} then a valid SourceHut username}

\item{pat}{see \code{\link[=sourcehut_pat]{sourcehut_pat()}}}
}
\description{
Retrieve metadata about yourself or another SourceHut user
}
\references{
\url{https://man.sr.ht/git.sr.ht/api.md}
}
