% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_audit}
\alias{meta_audit}
\title{Retrieve your audit log}
\usage{
meta_audit(pat = sourcehut_pat())
}
\arguments{
\item{pat}{see \code{\link[=sourcehut_pat]{sourcehut_pat()}}}
}
\description{
Retrieve your audit log
}
