% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_profile}
\alias{meta_profile}
\alias{meta_pgp_keys}
\title{Retrieve metadata about yourself}
\usage{
meta_profile(pat = sourcehut_pat())

meta_pgp_keys(id = NULL, next_id = 1, pat = sourcehut_pat())
}
\arguments{
\item{pat}{see \code{\link[=sourcehut_pat]{sourcehut_pat()}}}

\item{id}{if not \code{NULL} then a valid key ID}

\item{next_id}{if there is a \code{next} field in the API response and you wish
to paginate through the results, feed the value of \code{next} to
the \code{next_id} parameter.}
}
\description{
Retrieve metadata about yourself

Retrieve your PGP keys
}
\references{
\url{https://man.sr.ht/meta.sr.ht/user-api.md}
}
