% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcehut-package.R
\docType{package}
\name{sourcehut}
\alias{sourcehut}
\alias{sourcehut-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/sourcehut}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/sourcehut/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
