% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcehut-pat.R
\name{sourcehut_pat}
\alias{sourcehut_pat}
\title{Get or set SourceHut Personal Access Token}
\usage{
sourcehut_pat(force = FALSE)
}
\arguments{
\item{force}{force setting a new SourceHut Personal Access Token}
}
\value{
atomic character vector containing the SourceHut Personal Access Token
}
\description{
Get or set SourceHut Personal Access Token
}
\references{
\url{https://meta.sr.ht/oauth/personal-token}
}
