% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcehut-user.R
\name{sourcehut_user}
\alias{sourcehut_user}
\title{Get or set SourceHut username value}
\usage{
sourcehut_user(force = FALSE)
}
\arguments{
\item{force}{force setting a new SourceHut username for the current environment?}
}
\value{
atomic character vector containing the SourceHut username (with tilde)
}
\description{
Get or set SourceHut username value
}
