% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-builds.R
\name{use_builds}
\alias{use_builds}
\title{Setup .build.yml for SourceHut builds}
\usage{
use_builds(username = sourcehut_user(), open = TRUE)
}
\arguments{
\item{username}{your SourecHut user id (with or without the leading tilde)}

\item{open}{if \code{TRUE}, open the resultant file}
}
\description{
Setup .build.yml for SourceHut builds
}
