% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-sourcehut-badge.R
\name{use_sourcehut_badge}
\alias{use_sourcehut_badge}
\title{Adds a SourceHut builds.sr.ht badge to your README}
\usage{
use_sourcehut_badge(package = ".", username = sourcehut_user())
}
\arguments{
\item{package}{path to package (assumes current dir)}

\item{username}{see \code{\link[=sourcehut_user]{sourcehut_user()}}}
}
\description{
See \code{\link[usethis:use_badge]{usethis::use_badge()}} for details on what happens under the hood.
}
