% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcehut-package.R
\docType{package}
\name{sourcehut}
\alias{sourcehut}
\alias{sourcehut-package}
\title{sourcehut}
\description{
SourceHut (\url{https://sourcehut.org/}) provides support for collaborative
development of anything that can can be managed by Git or Mecurial version control
systems. Tools are provided to query and ochestrate a wide variety of SourceHut servies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/sourcehut}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/sourcehut/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
